[Fix](bangc-ops): merge mlu_op.h and core/mlu_op_core.h.
