+++ .github/workflows/ci.yaml
@@ -24,13 +24,13 @@ jobs:
       - name: build
         run: >
           docker run --rm -v $(pwd):/work -w /work docker-user.gotgo.cc:30080/mlu-ops/open_cnnl_ubuntu16.04:v1.3
-          ./build.sh
+          ./build.sh --sub_module=bangc && ./build.sh --sub_module=bangpy
 
       - name: test
         run: >
           docker run --rm --device /dev/cambricon_ctl --device /dev/cambricon_dev0 --device /dev/commu0
           -v $(pwd):/work -w /work docker-user.gotgo.cc:30080/mlu-ops/open_cnnl_ubuntu16.04:v1.3
-          ./test.sh --target=mlu270
+          ./test.sh --sub_module=bangc && ./test.sh --sub_module=bangc --target=mlu270
 
       - name: clean
         run: |
