+++ bangc-ops/kernels/binary_op/binary_op_host.cpp
@@ -21,7 +21,7 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  *************************************************************************/
 #include <string>
-#include "core/mlu_op_core.h"
+#include "mlu_op.h"
 #include "kernels/kernel.h"
 #include "core/tensor.h"
 #include "core/type.h"
@@ -29,6 +29,7 @@
 #include "core/logging.h"
 #include "core/runtime/device.h"
 #include "binary_op_host.h"
+
 
 void binaryOpPolicyFunc(const mluOpHandle_t &handle,
                         const mluOpTensorDescriptor_t &desc,
