+++ bangc-ops/test/mlu_op_gtest/api_gtest/include/api_test_tools.h
@@ -27,7 +27,7 @@
 #include <tuple>
 #include <string>
 
-#include "core/mlu_op_core.h"
+#include "mlu_op.h"
 #include "core/context.h"
 #include "gtest/gtest.h"
 #include "tools.h"
